Converted into unimodule
works on Android

iOS should work

fixed some bugs

applied requested changes
