+++ android/expoview/build.gradle
@@ -167,6 +167,7 @@ dependencies {
   compileOnly 'host.exp.exponent:expo-background-fetch:1.0.0'
   compileOnly 'host.exp.exponent:expo-mail-composer:1.0.0'
   compileOnly 'host.exp.exponent:expo-sqlite:1.0.0'
+  compileOnly 'host.exp.exponent:expo-image-picker:1.0.0'
   END UNCOMMENT WHEN DISTRIBUTING */
 
   // WHEN_DISTRIBUTING_REMOVE_FROM_HERE
@@ -210,6 +211,7 @@ dependencies {
   api project(':expo-gl-cpp')
   api project(':expo-google-sign-in')
   api project(':expo-image-loader-interface')
+  api project(':expo-image-picker')
   api project(':expo-keep-awake')
   api project(':expo-linear-gradient')
   api project(':expo-local-authentication')
