+++ android/settings.gradle
@@ -56,6 +56,7 @@ def includeUniversalModule = { name ->
     'expo-gl-cpp',
     'expo-google-sign-in',
     'expo-image-loader-interface',
+    'expo-image-picker',
     'expo-keep-awake',
     'expo-linear-gradient',
     'expo-local-authentication',
