--- ios/Exponent/Versioned/Core/Api/EXImagePicker.h
@@ -1,11 +1,0 @@
-#import "EXScopedBridgeModule.h"
-#import <UIKit/UIKit.h>
-
-typedef NS_ENUM(NSInteger, RNImagePickerTarget) {
-  RNImagePickerTargetCamera = 1,
-  RNImagePickerTargetLibrarySingleImage,
-};
-
-@interface EXImagePicker : EXScopedBridgeModule <UINavigationControllerDelegate, UIImagePickerControllerDelegate>
-
-@end
