+++ packages/expo-image-picker/android/build.gradle
@@ -1,0 +1,77 @@
+
+buildscript {
+  repositories {
+    google()
+    jcenter()
+    mavenCentral()
+  }
+
+  dependencies {
+    classpath 'com.android.tools.build:gradle:3.1.3'
+  }
+}
+
+apply plugin: 'com.android.library'
+apply plugin: 'maven'
+
+group = 'host.exp.exponent'
+version = '1.0.0'
+
+// Upload android library to maven with javadoc and android sources
+configurations {
+  deployerJars
+}
+
+// Creating sources with comments
+task androidSourcesJar(type: Jar) {
+  classifier = 'sources'
+  from android.sourceSets.main.java.srcDirs
+}
+
+// Put the androidSources and javadoc to the artifacts
+artifacts {
+  archives androidSourcesJar
+}
+
+uploadArchives {
+  repositories {
+    mavenDeployer {
+      configuration = configurations.deployerJars
+      repository(url: mavenLocal().url)
+    }
+  }
+}
+
+android {
+  compileSdkVersion 26
+
+  defaultConfig {
+    minSdkVersion 21
+    targetSdkVersion 26
+    versionCode 1
+    versionName "1.0.0"
+  }
+  lintOptions {
+    abortOnError false
+  }
+}
+
+if (new File(rootProject.projectDir.parentFile, 'pubspec.yaml').exists()) {
+  apply from: project(":expo_core").file("../expo-core.gradle")
+} else if (new File(rootProject.projectDir.parentFile, 'package.json').exists()) {
+  apply from: project(":expo-core").file("../expo-core.gradle")
+} else {
+  throw new GradleException(
+      "'expo-core.gradle' was not found in the usual Flutter or React Native dependency locations. " +
+          "This package can only be used in such projects. Are you sure you've installed the dependencies properly?")
+}
+
+dependencies {
+  expendency "expo-core"
+  expendency "expo-permissions-interface"
+  api 'com.theartofdev.edmodo:android-image-cropper:2.7.0'
+  api "com.android.support:support-v4:27.1.0"
+  api 'com.facebook.fresco:fresco:1.10.0'
+  api 'commons-codec:commons-codec:1.10'
+  api 'commons-io:commons-io:2.6'
+}
