+++ packages/expo-image-picker/android/src/main/AndroidManifest.xml
@@ -1,0 +1,5 @@
+
+<manifest package="expo.modules.imagepicker">
+
+</manifest>
+
