+++ packages/expo-image-picker/android/src/main/java/expo/modules/imagepicker/ImagePickerPackage.java
@@ -1,0 +1,16 @@
+package expo.modules.imagepicker;
+
+import android.content.Context;
+
+import java.util.Collections;
+import java.util.List;
+
+import expo.core.BasePackage;
+import expo.core.ExportedModule;
+
+public class ImagePickerPackage extends BasePackage {
+  @Override
+  public List<ExportedModule> createExportedModules(Context context) {
+    return Collections.singletonList((ExportedModule) new ImagePickerModule(context));
+  }
+}
