+++ packages/expo-image-picker/ios/EXImagePicker.podspec
@@ -1,0 +1,22 @@
+require 'json'
+
+package = JSON.parse(File.read(File.join(__dir__, '..', 'package.json')))
+
+Pod::Spec.new do |s|
+  s.name           = 'EXImagePicker'
+  s.version        = package['version']
+  s.summary        = package['description']
+  s.description    = package['description']
+  s.license        = package['license']
+  s.author         = package['author']
+  s.homepage       = package['homepage']
+  s.platform       = :ios, '10.0'
+  s.source         = { git: 'https://github.com/expo/expo.git' }
+  s.source_files   = 'EXImagePicker/**/*.{h,m}'
+  s.preserve_paths = 'EXImagePicker/**/*.{h,m}'
+  s.requires_arc   = true
+
+  s.dependency 'EXCore'
+  s.dependency 'EXFileSystemInterface'
+
+end
