+++ packages/expo-image-picker/src/ImagePicker/ExponentImagePicker.ts
@@ -1,0 +1,2 @@
+import { NativeModulesProxy } from 'expo-core';
+export default NativeModulesProxy.ExponentImagePicker;
