+++ packages/expo/package.json
@@ -85,6 +85,7 @@
     "expo-gl-cpp": "~2.0.0",
     "expo-google-sign-in": "~2.0.0",
     "expo-image-loader-interface": "~2.0.0",
+    "expo-image-picker": "~1.0.0",
     "expo-keep-awake": "~1.0.0",
     "expo-linear-gradient": "~1.0.0",
     "expo-local-authentication": "~2.0.0",
