+++ packages/expo/src/Expo.ts
@@ -25,7 +25,6 @@ import * as Calendar from './Calendar';
 import * as Google from './Google/Google';
 import * as Haptic from './Haptic/Haptic';
 import * as ImageManipulator from './ImageManipulator/ImageManipulator';
-import * as ImagePicker from './ImagePicker/ImagePicker';
 import * as IntentLauncherAndroid from './IntentLauncherAndroid/IntentLauncherAndroid';
 import * as ScreenOrientation from './ScreenOrientation/ScreenOrientation';
 import * as StoreReview from './StoreReview/StoreReview';
@@ -53,6 +52,8 @@ export { FileSystem } from 'expo-file-system';
 export { Font };
 export { GLView } from 'expo-gl';
 export { GoogleSignIn } from 'expo-google-sign-in';
+import * as ImagePicker from 'expo-image-picker';
+export { ImagePicker };
 export { LocalAuthentication } from 'expo-local-authentication';
 export { Localization } from 'expo-localization';
 export { Location };
@@ -78,7 +79,6 @@ export { Google };
 export { Haptic };
 export { default as Icon } from './Icon';
 export { ImageManipulator };
-export { ImagePicker };
 export { IntentLauncherAndroid };
 export { default as KeepAwake, activate, deactivate } from 'expo-keep-awake';
 export { default as Linking } from './Linking/Linking';
