--- packages/expo/src/ImagePicker/ExponentImagePicker.ts
@@ -1,3 +1,0 @@
-import { NativeModules } from 'react-native';
-
-export default NativeModules.ExponentImagePicker;
